+++ tests/dynamic_spec.rs
@@ -35,6 +35,31 @@ fn can_create_from_size_and_default() {
 }
 
 #[test]
+fn can_create_from_row_major_iter() {
+    let grid = Grid::new_from_iter(2, [1, 2, 3, 4]).unwrap();
+    assert_eq!(grid.get(0, 0), Some(&1));
+    assert_eq!(grid.get(1, 0), Some(&2));
+    assert_eq!(grid.get(0, 1), Some(&3));
+    assert_eq!(grid.get(1, 1), Some(&4));
+}
+
+#[test]
+fn cannot_create_from_invalid_row_major_iter() {
+    let _: IncompatibleRowSize = Grid::new_from_iter(2, [1, 2, 3, 4, 5]).unwrap_err();
+}
+
+#[test]
+fn cannot_create_from_invalid_row_major_iter_wth_invalid_width() {
+    let _: IncompatibleRowSize = Grid::new_from_iter(0, [1]).unwrap_err();
+}
+
+#[test]
+fn can_create_empty_grid_from_iter() {
+    let grid = Grid::<i32>::new_from_iter(0, []).unwrap();
+    assert!(grid.is_empty());
+}
+
+#[test]
 fn can_mutate_cell() {
     let mut grid = Grid::new_with_default(2, 2);
     assert_eq!(grid.get(1, 0), Some(&0));